refactor: introduce constants, fix order and tests

+++ src/main/kotlin/com/linxhealth/controller/dto/AppointmentDTO.kt
@@ -3,6 +3,7 @@ package com.linxhealth.controller.dto
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.linxhealth.model.AppointmentStatus
 import io.micronaut.serde.annotation.Serdeable
+import java.math.BigDecimal
 
 @Serdeable
 data class AppointmentRequest(
@@ -26,12 +27,12 @@ data class AppointmentResponse(
 
 @Serdeable
 data class BillResponse(
-    val fee: Double,
-    @field:JsonProperty("discount_percentage") val discountPercentage: Double,
-    @field:JsonProperty("discount_amount") val discountAmount: Double,
-    @field:JsonProperty("amount_after_discount") val amountAfterDiscount: Double,
-    @field:JsonProperty("tax_amount") val taxAmount: Double,
-    @field:JsonProperty("total_amount") val totalAmount: Double,
-    @field:JsonProperty("insurance_amount") val amountCoveredByInsurance: Double,
-    @field:JsonProperty("co_pay_amount") val coPayAmount: Double
+    val fee: BigDecimal,
+    @field:JsonProperty("discount_percentage") val discountPercentage: BigDecimal,
+    @field:JsonProperty("discount_amount") val discountAmount: BigDecimal,
+    @field:JsonProperty("amount_after_discount") val amountAfterDiscount: BigDecimal,
+    @field:JsonProperty("tax_amount") val taxAmount: BigDecimal,
+    @field:JsonProperty("total_amount") val totalAmount: BigDecimal,
+    @field:JsonProperty("insurance_amount") val amountCoveredByInsurance: BigDecimal,
+    @field:JsonProperty("co_pay_amount") val coPayAmount: BigDecimal
 )
